+++ frontend/src/UI/components/Navigation/index.ts
@@ -1,1 +1,2 @@
 export * from "./NavigationBar";
+export * from "./SideMenu";
