+++ frontend/src/business/overmind/Review/actions.ts
@@ -1,13 +1,15 @@
 import { pipe } from "overmind";
-import { getAuthors } from "../Author/operators";
-import { getBooks } from "../Book/operators";
+import { getAuthors, shouldLoadAuthors } from "../Author/operators";
+import { getBooks, shouldLoadBooks } from "../Book/operators";
 import * as o from "./operators";
 
 export const showReviewsPage = pipe(o.setReviewsPage(), o.getReviews());
 
 export const showAddReviewPage = pipe(
   o.setAddReviewPage(),
-  o.shouldLoadAuthorsAndBooks(),
+  shouldLoadAuthors(),
+  shouldLoadBooks,
   getAuthors(),
+  // should be fixed with an api call, with author id
   getBooks()
 );
